SimTK installation.

+++ gforge/common/frs/FRSRelease.class.php
@@ -5,6 +5,7 @@
  * Copyright 2002, Tim Perdue/GForge, LLC
  * Copyright 2009, Roland Mas
  * Copyright (C) 2012 Alain Peyrat - Alcatel-Lucent
+ * Copyright 2016, Henry Kwong, Tod Hing - SimTK Team
  *
  * This file is part of FusionForge. FusionForge is free software;
  * you can redistribute it and/or modify it under the terms of the
@@ -114,9 +115,10 @@ class FRSRelease extends Error {
 	 * @param	string	$changes	The change log for the release.
 	 * @param	int	$preformatted	Whether the notes/log are preformatted with \n chars (1) true (0) false.
 	 * @param	int	$release_date	The unix date of the release.
+	 * @param	int	$status_id	Active/Hidden status of the release.
 	 * @return	boolean	success.
 	 */
-	function create($name,$notes,$changes,$preformatted,$release_date=false) {
+	function create($name,$notes,$changes,$preformatted,$release_date=false, $status_id=1, $release_desc='') {
 		if (strlen($name) < 3) {
 			$this->setError(_('FRSPackage Name Must Be At Least 3 Characters'));
 			return false;
@@ -145,15 +147,22 @@ class FRSRelease extends Error {
 		}
 
 		db_begin();
-		$result=db_query_params ('INSERT INTO frs_release(package_id,notes,changes,preformatted,name,release_date,released_by,status_id) VALUES ($1,$2,$3,$4,$5,$6,$7,$8)',
-					 array ($this->FRSPackage->getId(),
-						htmlspecialchars($notes),
-						htmlspecialchars($changes),
-						$preformatted,
-						htmlspecialchars($name),
-						$release_date,
-						user_getid(),
-						1)) ;
+		$result = db_query_params('INSERT INTO frs_release (' .
+			'package_id, notes, changes, preformatted, name, ' .
+			'release_date, released_by, status_id, simtk_description) VALUES ' .
+			'($1,$2,$3,$4,$5,$6,$7,$8,$9)',
+			array(
+				$this->FRSPackage->getId(),
+				htmlspecialchars($notes),
+				htmlspecialchars($changes),
+				$preformatted,
+				htmlspecialchars($name),
+				$release_date,
+				user_getid(),
+				$status_id,
+				htmlspecialchars($release_desc)
+			)
+		);
 		if (!$result) {
 			$this->setError(_('Error Adding Release: ').db_error());
 			db_rollback();
@@ -169,8 +178,13 @@ class FRSRelease extends Error {
 				@mkdir($newdirlocation);
 			}
 			db_commit();
-			return true;
 		}
+
+		// Release created.
+		// Send notice to users monitoring this package.
+		$this->sendNotice();
+
+		return true;
 	}
 
 	/**
@@ -238,6 +252,15 @@ class FRSRelease extends Error {
 	}
 
 	/**
+	 * getDesc - get the description of this release.
+	 *
+	 * @return	int	The description.
+	 */
+	function getDesc() {
+		return $this->data_array['simtk_description'];
+	}
+
+	/**
 	 * getNotes - get the release notes of this release.
 	 *
 	 * @return	string	The release notes.
@@ -278,7 +301,7 @@ class FRSRelease extends Error {
 	 *
 	 * @return	boolean	success.
 	 */
-	function sendNotice() {
+	function sendNotice($action=false) {
 		$arr =& $this->FRSPackage->getMonitorIDs();
 
 		$date = date('Y-m-d H:i',time());
@@ -286,29 +309,42 @@ class FRSRelease extends Error {
 		$subject = sprintf (_('[%1$s Release] %2$s'),
 					$this->FRSPackage->Group->getUnixName(),
 					$this->FRSPackage->getName());
-		$text = sprintf(_('Project %1$s (%2$s) has released a new version of package “%3$s”.'),
-										$this->FRSPackage->Group->getPublicName(),
-										$this->FRSPackage->Group->getUnixName(),
-										$this->FRSPackage->getName())
-							. "\n\n"
-							. _('Release Notes')._(':')
-							. "\n\n"
-							. $this->getNotes()
-							. _('Change Log')._(':')
-							. "\n\n"
-							. $this->getChanges()
-							. "\n\n"
-							. _('You can download it by following this link')._(':')
-							. "\n\n"
-							. util_make_url ("/frs/?group_id=". $this->FRSPackage->Group->getID() ."&release_id=". $this->getID())
-							. "\n\n"
-							. sprintf(_('You receive this email because you requested to be notified when new '
-										. 'versions of this package were released. If you don\'t wish to be '
-										. 'notified in the future, please login to %s and click this link:'),
-										forge_get_config('forge_name'))
-							. "\n\n"
-							. util_make_url ("/frs/monitor.php?filemodule_id=".$this->FRSPackage->getID()."&group_id=".$this->FRSPackage->Group->getID()."&stop=1");
-//		$text = util_line_wrap($text);
+
+		$formatStr = 'Project %1$s (%2$s) has released a new version of package "%3$s".';
+		if ($action == 'UPDATE_RELEASE') {
+			$formatStr = 'Project %1$s (%2$s) has updated release in package "%3$s".';
+		}
+		else if ($action == 'DELETE_RELEASE') {
+			$formatStr = 'Project %1$s (%2$s) has deleted a release in package "%3$s".';
+		}
+		$content = sprintf($formatStr,
+			$this->FRSPackage->Group->getPublicName(),
+			$this->FRSPackage->Group->getUnixName(),
+			$this->FRSPackage->getName());
+
+		$text = $content . 
+			"\n\n" . 
+			_('Release Notes')._(':') . 
+			"\n\n" . 
+			$this->getNotes() . 
+			_('Change Log')._(':') . 
+			"\n\n" . 
+			$this->getChanges() . 
+			"\n\n" . 
+			_('You can download it by following this link')._(':') . 
+			"\n\n" . 
+			util_make_url("/frs/?group_id=". 
+				$this->FRSPackage->Group->getID() .
+				"&release_id=". $this->getID()) . 
+			"\n\n" . 
+			sprintf(_('You receive this email because you requested to be notified when new ' . 
+				'versions of this package were released. If you don\'t wish to be ' . 
+				'notified in the future, please login to %s and click this link:'), 
+				forge_get_config('forge_name')) . 
+			"\n\n" . 
+			util_make_url("/frs/monitor.php?filemodule_id=".
+				$this->FRSPackage->getID() . "&group_id=" .
+				$this->FRSPackage->Group->getID() . "&stop=1");
 		if (count($arr)) {
 			util_handle_message(array_unique($arr),$subject,$text);
 		}
@@ -333,8 +369,21 @@ class FRSRelease extends Error {
 	function &getFiles() {
 		if (!is_array($this->release_files) || count($this->release_files) < 1) {
 			$this->release_files=array();
+/*
 			$res = db_query_params ('SELECT * FROM frs_file_vw WHERE release_id=$1',
 						array ($this->getID())) ;
+*/
+			// Need Simtk 1.0 data here. See fetchData() in FRSFile.class.php.
+			$res = db_query_params ('SELECT * FROM frs_file_vw as ffv ' .
+				'JOIN (SELECT file_id, simtk_description, ' .
+				'simtk_collect_data, simtk_use_mail_list, simtk_group_list_id, ' .
+				'simtk_filetype, simtk_filelocation, ' .
+				'simtk_show_notes, simtk_show_agreement, ' .
+				'simtk_rank, simtk_filename_header ' .
+				'FROM frs_file) AS ff ' .
+				'ON ffv.file_id=ff.file_id ' .
+				'WHERE release_id=$1',
+				array ($this->getID())) ;
 			while ($arr = db_fetch_array($res)) {
 				$this->release_files[]=$this->newFRSFile($arr['file_id'],$arr);
 			}
@@ -360,7 +409,7 @@ class FRSRelease extends Error {
 		}
 		$f =& $this->getFiles();
 		for ($i=0; $i<count($f); $i++) {
-			if (!is_object($f[$i]) || $f[$i]->isError() || !$f[$i]->delete()) {
+			if (!is_object($f[$i]) || $f[$i]->isError() || !$f[$i]->delete(true)) {
 				$this->setError('File Error: '.$f[$i]->getName().':'.$f[$i]->getErrorMessage());
 				return false;
 			}
@@ -375,11 +424,18 @@ class FRSRelease extends Error {
 			$this->setError(_('Release delete error: trying to delete root dir'));
 			return false;
 		}
-		rmdir($dir);
+		if (is_dir($dir)) {
+			rmdir($dir);
+		}
 
 		db_query_params ('DELETE FROM frs_release WHERE release_id=$1 AND package_id=$2',
 				 array ($this->getID(),
 					$this->FRSPackage->getID())) ;
+
+		// This release deleted.
+		// Send notice to users monitoring this package.
+		$this->sendNotice('DELETE_RELEASE');
+
 		return true;
 	}
 
@@ -392,9 +448,10 @@ class FRSRelease extends Error {
 	 * @param	string	The change log for the release.
 	 * @param	int	Whether the notes/log are preformatted with \n chars (1) true (0) false.
 	 * @param	int	The unix date of the release.
+	 * @param	string	The description of the release.
 	 * @return	boolean success.
 	 */
-	function update($status, $name, $notes, $changes, $preformatted, $release_date) {
+	function update($status, $name, $notes, $changes, $preformatted, $release_date, $release_desc) {
 		if (strlen($name) < 3) {
 			$this->setError(_('FRSPackage Name Must Be At Least 3 Characters'));
 			return false;
@@ -412,27 +469,36 @@ class FRSRelease extends Error {
 		}
 
 		if($this->getName() != htmlspecialchars($name)) {
-			$res = db_query_params ('SELECT * FROM frs_release WHERE package_id=$1 AND name=$2',
-						array ($this->FRSPackage->getID(),
-							   htmlspecialchars($name))) ;
+			$res = db_query_params('SELECT * FROM frs_release WHERE package_id=$1 AND name=$2',
+				array(
+					$this->FRSPackage->getID(), 
+					htmlspecialchars($name)
+				)
+			);
 			if (db_numrows($res)) {
 				$this->setError(_('Error On Update: Name Already Exists'));
 				return false;
 			}
 		}
 		db_begin();
-		$res = db_query_params ('UPDATE frs_release SET	name=$1,status_id=$2,notes=$3,
-			changes=$4,preformatted=$5,release_date=$6,released_by=$7
-			WHERE package_id=$8 AND release_id=$9',
-					array (htmlspecialchars($name),
-						   $status,
-						   htmlspecialchars($notes),
-						   htmlspecialchars($changes),
-						   $preformatted,
-						   $release_date,
-						   user_getid(),
-						   $this->FRSPackage->getID(),
-						   $this->getID())) ;
+		$res = db_query_params('UPDATE frs_release SET ' .
+			'name=$1, status_id=$2, notes=$3, changes=$4, ' .
+			'preformatted=$5, release_date=$6, ' .
+			'released_by=$7, simtk_description=$10 ' .
+			'WHERE package_id=$8 AND release_id=$9',
+			array(
+				htmlspecialchars($name),
+				$status,
+				htmlspecialchars($notes),
+				htmlspecialchars($changes),
+				$preformatted,
+				$release_date,
+				user_getid(),
+				$this->FRSPackage->getID(),
+				$this->getID(),
+				htmlspecialchars($release_desc)
+			)
+		);
 
 		if (!$res || db_affected_rows($res) < 1) {
 			$this->setError(sprintf(_('Error On Update: %s'), db_error()));
@@ -464,10 +530,76 @@ class FRSRelease extends Error {
 			}
 		}
 		db_commit();
+
+		// Release updated.
+		// Send notice to users monitoring this package.
+		$this->sendNotice('UPDATE_RELEASE');
+
 		$this->FRSPackage->createNewestReleaseFilesAsZip();
 		return true;
 	}
 
+	/**
+	 * arrangeFiles - Arrange files of a release in the database.
+	 *
+	 * @param	int	The file id.
+	 * @param	int	The rank of the file.
+	 * @param	string	The header of the file
+	 * @return	boolean success.
+	 */
+	function arrangeFiles($fileId, $rank, $header) {
+
+		if (!forge_check_perm('frs', $this->FRSPackage->Group->getID(), 'write')) {
+			$this->setPermissionDeniedError();
+			return false;
+		}
+
+		db_begin();
+
+		$res = db_query_params('UPDATE frs_file SET simtk_rank=$1, simtk_filename_header=$2 
+			WHERE file_id=$3 AND release_id=$4',
+			array(
+				$rank, 
+				htmlspecialchars($header), 
+				$fileId, $this->getID()
+			)
+		);
+
+		if (!$res || db_affected_rows($res) < 1) {
+			$this->setError(sprintf(_('Error On Update: %s'), db_error()));
+			db_rollback();
+			return false;
+		}
+
+		db_commit();
+
+		return true;
+	}
+	
+	function setDoi($doi=1) {
+	
+	    db_begin();
+		$res = db_query_params('UPDATE frs_release SET ' .
+			'doi=$1 ' .
+			'WHERE package_id=$2 AND release_id=$3',
+			array(
+				$doi,
+				$this->FRSPackage->getID(),
+				$this->getID()
+			)
+		);
+
+		if (!$res || db_affected_rows($res) < 1) {
+			$this->setError(sprintf(_('Error On Update: %s'), db_error()));
+			db_rollback();
+			return false;
+		}
+        
+		db_commit();
+
+		return true;
+	}
+	
 }
 
 // Local Variables:
